Fixes several issues with counting Test Cases:
1)	Data Issues: In two situations possible, the Test Case Count  was obviously inaccurate: 
a.	Test Cases Passing count was set greater than the Total Test Case Count
b.	The overall TA test result was set to “Pass” but the Test Cases Passing is less than Test Cases Failing
A question mark is now placed in front of the values in the TC Pass/Total column when either of these cases exist. This flags for you it so you can make an update.
2)	Pass/Fail Highlighting. The values in the  TC Pass/Total column now follow the test result value: green for a pass and red for a fail. It is standard black if no result has been recorded.
3)	Test Case Pass Rate: The Test Case pass rate was incorrectly counting the Test Cases in the following situations:
a.	The Test Case Passing  number was only counted when the entire TA had a value of “Pass”. Now it counts whatever the last total passing is recorded in the test results notes, that is, whatever was last recorded as Test Case followed by a semicolon in the test results .
b.	The report will now catch previously entered TC Passing numbers in the notes that have don’t have a semicolon – those with a space or colon after them.