KRACOEUS-6803: Created new method KraWorkflowService.isCurrentNode which internally uses WorkflowDocument method getCurrentNodeNames to obtain both active and terminal route node names.  This new method had to be created because we needed to be able find the nodeName passed in when it identifies a terminal node (i.e. the document has completed its routing so that it is in processed or final status).  Method KraWorkflowService.isDocumentOnNode internally uses WorkflowDocument method getNodeNames which returns null when input parameter nodeName identifies a terminal node necessitating this change.
Adjusted code to also set Committee drop down on the Assign to Committee and Schedule action sub-panel to the value that was selected by the user when they completed the Request for Suspension action.  Could not set the Schedule on Assign to Committee and Schedule sub-panel to a value as there is no corresponding drop down on the Request for Suspension sub-panel.