KRACOEUS-6803: These changes are for IRB: Code was modified to only allow PI to see the Withdraw Suspension action after a request for suspension.  Code was modified to allow the admin to see the additional actions of Assign to Reviewers, Assign to Agenda (once user selects a schedule date), and Modify Submission Request  after a request for suspension is performed.    [NOTE:]  The Request for Suspension action sets the submission review type to Full. This submission review type value REQUIRES a schedule date before the Assign to Reviewers action will show.
All submission types values were coded to be available for the Modify Submission Request action after a user performs a Request for Suspension action.  

Refactored IRB method findSubmission as it had the same code duplicated in four specific IRB authorizers (ProtocolAssignReviewersAuthorizer, ProtocolAssignReviewersUnavailableAuthorizer, ProtocolAssignToAgendaAuthorizer, ProtocolAssignToAgendaUnavailableAuthorizer).  This was done so the so that this code would not need to be duplicated six more times in other specific IRB authorizers. Modeled the findSubmission methods placed in IRB base class ProtocolAuthorizer after ones in ProtocolAssignCmtSchedServiceImpl.  This modeling was done due to visibility to these service layer methods and concerns about  refactoring the service layer at this time.

To deal with need to specify user role of Admin for test cases the following was done in class ProtocolAuthorizerTestBase:
1)	New method signature was created that included roleConstant and rolePermission.  This new signature replaced the existing signature and was used for the test cases that now require a role.
2)	Existing method signature was made to call the new method signature passing the role of RoleConstants.PROTOCOL_AGGREGATOR and true.  This is to allow existing calls to not need to be recoded.