KRACOEUS-6803: These changes are for IACUC:
Code was modified to allow PI to only see the Withdraw Suspension action after a request for suspension. Code was modified to allow the admin role to see the additional actions of Assign to Reviewers, Assign to Agenda (once user selects a schedule date), and Modify Submission Request after a request for suspension is performed. [NOTE:] The Request for Suspension action sets the submission review type to Full. This submission review type value REQUIRES a schedule date before the Assign to Reviewers action will show.

All submission types values were coded to be available for the Modify Submission Request action after a user performs a Request for Suspension action.