KRACOEUS-6652: IRB changes:  Added code to the modify submission bean prepare view method so that the bean presented the data from the current submission. Added code to model in IRB what was already in place for IACUC to have the modify submission bean refresh itself after a new submission element was created. Added a call to the notify protocol service implementation to tell the modify submission bean to refresh itself due to a new submission item being created.  Adjusted code to retain the error messages correctly set by the business rules so that they would not be cleared and replaced by a generic message allowing the user to know the real reason for the error so they can fix the problem.
IACUC changes: Adjusted the modify submission bean code so that all of the data it holds is for the current submission and adjusted the code so the reviewers only get populated once.  Added a call to the notify protocol service implementation to tell the modify submission bean to refresh itself due to a new submission item being created. Adjusted submission type values finder to show the Notify Iacuc description instead of the code for a particular case found during testing.